requirements.txt
appdirs==1.4.4
asgiref==3.2.7
boto3==1.14.45
botocore==1.17.45
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
cryptography==2.9.2
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.0.8
django-appconf==1.0.4
django-avatar==5.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-environ==0.4.5
django-environment==0.3.8
django-js-asset==1.2.2
django-registration==3.1
django-storages==1.9.1
django-summernote==0.8.11.6
django-taggit==1.3.0
django-tinymce==3.0.2
docutils==0.15.2
ecdsa==0.15
filelock==3.0.12
idna==2.10
jmespath==0.10.0
Markdown==3.2.2
oauthlib==3.1.0
pbr==5.4.5
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-jose==3.1.0
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==3.2.0
urllib3==1.25.9
virtualenv==20.0.30
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
